split shop api and resourses into multiple files; misc cleanups
