+++ examples/fetch_old_results.py
@@ -1,4 +1,7 @@
 import tempfile
+from pprint import pprint
+
+import yaml
 
 from cognite.powerops import PowerOpsClient
 from cognite.powerops.logger import configure_debug_logging
@@ -9,7 +12,7 @@ p = PowerOpsClient()
 
 run = p.shop.runs.retrieve(external_id="POWEROPS_SHOP_RUN_ede8c4c0-18b1-41b1-ae40-ea20e037645c")
 
-print(run.is_complete)
+print(run.in_progress)
 print(run.status)
 
 res = run.get_results()
@@ -18,6 +21,8 @@ tmp_dir = tempfile.mkdtemp(prefix="power-ops-sdk-usage")
 print(res.post_run.save(tmp_dir))
 
 print(res.post_run["model.objective.average_objective.solver_status"])
+pprint(res.post_run["model.objective.average_objective"])
+print(yaml.dump(res.post_run["model.objective.average_objective"]))
 # res.plot(
 #     "model.creek_intake.Golebiowski_intake.net_head",
 # )
