+++ examples/usage.py
@@ -34,7 +34,7 @@ sample_case = powerops.shop.cases.create(model_template=sample_model, commands=[
 
 sample_run = powerops.shop.runs.trigger(case=sample_case)
 
-print(sample_run.is_complete)
+print(sample_run.in_progress)
 
 
 sample_run_results = powerops.shop.runs.trigger(case=sample_case).get_results()
